Implement reauthentication, access token refresh
This allows a special state to the authenticate() call so you
can get a new access token and have it stored in the ember-simple-auth
authenticated data (in particular, where ember-data looks for it).

Note this is currently not called anywhere -- soon I'll have
a place in the cognito service where you can enable a recurring timer.